requirements.txt
opencv-python
pytesseract
pillow
gradio
numpy
pandas
pdf2image

